MAINT refactored utils
