+++ setup.py
@@ -5,8 +5,8 @@ long_description = "An easy to use Python 3 Pandas Extension with 115+ Technical
 
 setup(
     name ="pandas_ta",
-    packages =["pandas_ta", "pandas_ta.candles", "pandas_ta.momentum", "pandas_ta.overlap", "pandas_ta.performance", "pandas_ta.statistics", "pandas_ta.trend", "pandas_ta.volatility", "pandas_ta.volume"],
-    version =".".join(("0", "2", "14b")),
+    packages =["pandas_ta", "pandas_ta.candles", "pandas_ta.momentum", "pandas_ta.overlap", "pandas_ta.performance", "pandas_ta.statistics", "pandas_ta.trend", "pandas_ta.utils", "pandas_ta.volatility", "pandas_ta.volume"],
+    version =".".join(("0", "2", "15b")),
     description =long_description,
     long_description =long_description,
     author ="Kevin Johnson",
